+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Don't emit comments with `lvl system sshconfig` to avoid parsing issues on certain OpenSSH versions.
 * Fix `lvl system describe` failing due to changes in API.
+* New "id" output mode to print only IDs of values returned. This is for ease of use in scripts.
 
 ## 1.4.0